requirements.txt
SimpleITK
torch
numpy==1.26.4
timm
torchvision
Pillow
opencv-python-headless
typing-extensions